Restructed code to make it more customizable on feature creation.
