--- Model/src/features/VerboseValueNullBinary.py
@@ -1,61 +1,0 @@
-from src.features.FeatureETLEngineering import FeatureETLEngineering
-from collections import Counter
-import re
-import pandas as pd
-
-
-class VerboseValueNullBinary(FeatureETLEngineering):
-    def __init__(self, col_series, threshold, most_common):
-        """Instantiate Verbose Feature."""
-        super().__init__(col_series)
-        self.most_common = most_common
-        self.threshold = threshold
-
-    def run_etl(self):
-        """Run ETL of the column."""
-        self._replace_nans()
-        self._add_is_null_column_df()
-        self._string_to_list()
-        self._word_count_catergorical()
-        self._build_feature_etl_df()
-
-    def _string_to_list(self):
-        """Clean up strings and transform them into list of strings."""
-
-        def _remove_punctuation_make_list(row):
-            try:
-                clean = re.sub(r'[^\w\s]', '', row.lower()).split(' ')
-                return clean
-
-            except:
-                # Run it and see if there are many rows with parse error
-                return ['none']
-
-        # Lower case string, Remove Punctuation, and Split on ' '
-        self.col_etl = self.col_etl.apply(
-            lambda x: _remove_punctuation_make_list(x))
-
-    def _word_count_catergorical(self):
-        """Create Catergorical columns based on word count thresholds and
-        whether or not the current observation contains said word."""
-
-        # Count all instances of a word
-        results = Counter()
-        self.col_etl.apply(results.update)
-
-        # Make list of words over threshold if most_common = True
-        if self.most_common:
-            word_feats = [(x) for (x, y) in results.most_common()
-                          if y > self.threshold]
-
-        # Else Make list of words below threshold if most_common = False
-        else:
-            word_feats = [(x) for (x, y) in results.most_common()
-                          if y <= self.threshold]
-
-        # For Each word in list, make column and assign Binary value
-        for word in word_feats:
-
-            # Return binary, whether iteration is in list of strings
-            self.df_etl[f'{self.col_etl.name}_cat_{word}'] = self.col_etl.apply(
-                lambda x: 1 if word in x else 0)
