+++ Model/src/features/categorical/CategoricalFeatureCreator.py
@@ -6,16 +6,16 @@ import pandas as pd
 class CategoricalFeatureCreator():
     def __init__(self, dataset, nominal_cat_columns=[]):
         """Create Categorical Nominal Features."""
-
         # Inputs
         self.dataset = dataset
         self.nominal_cat_columns = nominal_cat_columns
 
         # Outputs
         self.cat_custom_nominals = []
-        self.df_model = pd.DataFrame()
+        self.df_etl = pd.DataFrame()
 
     def _create_categorical_features(self):
+        """ETL our Categorical Features."""
         # self._create_custom_categorical_nominal_features()
         self._create_categorical_nominal_features()
     
@@ -27,8 +27,8 @@ class CategoricalFeatureCreator():
         categorical = CategoricalValueNullBinaryNominalCustom_area(
             self.dataset['area'], name='area').run_etl()
         self.cat_custom_nominals = self.cat_custom_nominals + ['area']
-        self.df_model = pd.concat(
-                [self.df_model, categorical.df_etl], axis=1)
+        self.df_etl = pd.concat(
+                [self.df_etl, categorical.df_etl], axis=1)
 
     def _create_categorical_nominal_features(self):
         """Create categorical nominal features and add to Model Input DF."""
@@ -38,11 +38,9 @@ class CategoricalFeatureCreator():
             set(self.cat_custom_nominals))
 
         # Loop through each column and create categorical feature
-        print(self.nominal_cat_columns)
         for column in self.nominal_cat_columns:
-            print(self.dataset[column])
-            categorical2 = CategoricalValueNullBinaryNominal(self.dataset[column])
-            categorical2.run_etl()
+            categorical= CategoricalValueNullBinaryNominal(self.dataset[column])
+            categorical.run_etl()
 
-            self.df_model = pd.concat(
-                [self.df_model, categorical2.df_etl], axis=1)
+            self.df_etl = pd.concat(
+                [self.df_etl, categorical.df_etl], axis=1)
